pyproject.toml

[tool.poetry]
name = "mattrizzo_devops"
version = "1.1.0"
description = "Repository containing all things related to CI and devops for any code I develop"
authors = ["Matt Rizzo <mattrizzo21@gmail.com>"]
license = "MIT License"
repository = "https://github.com/MatthewRizzo/mattrizzo_devops"

packages = [{include = "hooks"}]

[tool.poetry.scripts]
check-rust = "hooks.pre_push.check_rust:main"
check-python-typing = "hooks.pre_push.run_mypy:main"

[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "2.20.0"
GitPython = "^3.1.29"
pylint = "^2.15.4"
mypy = "^0.982"

[tool.poetry.dev-dependencies]
poetry = "1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
